+++ Backend/project/src/main/java/com/pac/gestoreeventi/profileManagement/ProfileRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProfileRepository extends JpaRepository<Profile,Long> {
-    Optional<Profile> findById(Integer idProfile);
+public interface ProfileRepository extends JpaRepository<Profile, Integer> {
+    Optional<Profile> findById(int idProfile);
     Optional<Profile> findByEmail(String email);
 }
