+++ Backend/project/src/main/java/com/pac/gestoreeventi/reservationManagement/ReservationManagementIF.java
@@ -3,5 +3,5 @@ package com.pac.gestoreeventi.reservationManagement;
 public interface ReservationManagementIF {
     void createReservation(ReservationDTO reservationDTO);
 
-    void deleteReservation(Long idReservation);
+    void deleteReservation(Integer idReservation);
 }
